# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/eagle/imports/libraries": {
      "get": {
        "operationId": "EagleImportController_listLibraries",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports": {
      "post": {
        "operationId": "EagleImportController_createRun",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEagleImportRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}": {
      "get": {
        "operationId": "EagleImportController_getRun",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/manifest/chunks": {
      "post": {
        "operationId": "EagleImportController_stageChunk",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EagleImportManifestChunkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/preflight": {
      "post": {
        "operationId": "EagleImportController_preflight",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/items": {
      "get": {
        "operationId": "EagleImportController_listItems",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/items/{itemId}/upload": {
      "post": {
        "operationId": "EagleImportController_initiateUpload",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/items/{itemId}/finish": {
      "post": {
        "operationId": "EagleImportController_finishItem",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishEagleImportItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
    "/api/eagle/imports/{runId}/cancel": {
      "post": {
        "operationId": "EagleImportController_cancel",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "tags": ["eagle-imports"]
      }
    },
...
    "/api/admin/eagle-processing/jobs/{jobId}/retry": {
      "post": {
        "operationId": "EagleProcessingController_retry",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["EagleProcessing"]
      }
    },
...
    "/api/admin/eagle-processing/settings": {
      "patch": {
        "operationId": "EagleProcessingController_settings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEagleProcessingSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["EagleProcessing"]
      }
    },
...
  },
  "info": {
    "title": "SekerEagle API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateEagleImportRunDto": {
        "type": "object",
        "properties": {}
      },
      "EagleImportManifestChunkDto": {
        "type": "object",
        "properties": {}
      },
      "FinishEagleImportItemDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          }
        },
        "required": ["assetId"]
      },
      "UpdateEagleProcessingSettingsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}